# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Loom API",
    "description": "Subtitle generation engine over loom_core.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/generate/ass": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "Generate Ass",
        "operationId": "generate_ass_generate_ass_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateAssRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateAssResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate/pgs": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "Generate Pgs",
        "operationId": "generate_pgs_generate_pgs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneratePgsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate/suggest-filename": {
      "post": {
        "tags": [
          "generate"
        ],
        "summary": "Suggest Filename",
        "operationId": "suggest_filename_generate_suggest_filename_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestFilenameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestFilenameResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BottomLayerStyle": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "fontname": {
            "type": "string",
            "title": "Fontname",
            "default": "Georgia"
          },
          "fontsize": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Fontsize",
            "default": 48
          },
          "bold": {
            "type": "boolean",
            "title": "Bold",
            "default": false
          },
          "italic": {
            "type": "boolean",
            "title": "Italic",
            "default": false
          },
          "primarycolor": {
            "type": "string",
            "title": "Primarycolor",
            "default": "#FFFFFF"
          },
          "primary_opacity": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Primary Opacity",
            "default": 100
          },
          "outlinecolor": {
            "type": "string",
            "title": "Outlinecolor",
            "default": "#000000"
          },
          "outline_opacity": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Outline Opacity",
            "default": 100
          },
          "backcolor": {
            "type": "string",
            "title": "Backcolor",
            "default": "#000000"
          },
          "back_opacity": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Back Opacity",
            "default": 0
          },
          "outline_enabled": {
            "type": "boolean",
            "title": "Outline Enabled",
            "default": true
          },
          "outline": {
            "type": "number",
            "minimum": 0.0,
            "title": "Outline",
            "default": 3.0
          },
          "shadow_enabled": {
            "type": "boolean",
            "title": "Shadow Enabled",
            "default": false
          },
          "shadow": {
            "type": "number",
            "minimum": 0.0,
            "title": "Shadow",
            "default": 1.5
          },
          "background_enabled": {
            "type": "boolean",
            "title": "Background Enabled",
            "default": false
          },
          "glow_enabled": {
            "type": "boolean",
            "title": "Glow Enabled",
            "default": false
          },
          "glow_radius": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Glow Radius",
            "default": 5
          },
          "glow_color_hex": {
            "type": "string",
            "title": "Glow Color Hex",
            "default": "#FFFF00"
          },
          "alignment": {
            "type": "integer",
            "maximum": 9.0,
            "minimum": 1.0,
            "title": "Alignment",
            "default": 2
          },
          "marginl": {
            "type": "integer",
            "title": "Marginl",
            "default": 10
          },
          "marginr": {
            "type": "integer",
            "title": "Marginr",
            "default": 10
          },
          "marginv": {
            "type": "integer",
            "title": "Marginv",
            "default": 40
          }
        },
        "type": "object",
        "title": "BottomLayerStyle"
      },
...
      "GenerateAssRequest": {
        "properties": {
          "native_file_id": {
            "type": "string",
            "title": "Native File Id"
          },
          "target_file_id": {
            "type": "string",
            "title": "Target File Id"
          },
          "target_lang_code": {
            "type": "string",
            "title": "Target Lang Code"
          },
          "styles": {
            "$ref": "#/components/schemas/StyleConfig"
          },
          "offsets": {
            "$ref": "#/components/schemas/TimingOffsets",
            "default": {
              "bottom_ms": 0,
              "top_ms": 0
            }
          },
          "source_resolution": {
            "$ref": "#/components/schemas/Resolution",
            "default": {
              "width": 1920,
              "height": 1080
            }
          },
          "output_resolution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Resolution"
              },
              {
                "type": "null"
              }
            ]
          },
          "include_annotations": {
            "type": "boolean",
            "title": "Include Annotations",
            "default": false
          },
          "opt_in_training": {
            "type": "boolean",
            "title": "Opt In Training",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "native_file_id",
          "target_file_id",
          "target_lang_code",
          "styles"
        ],
        "title": "GenerateAssRequest",
        "description": "Client \u2192 server request for ASS generation.\n\nFile handles cross the wire as IDs (returned by POST /files), not paths.\nThe server resolves IDs \u2192 on-disk paths just before calling the engine.\n\n``opt_in_training``: when true, the server is permitted to archive\n``(input_text, style_config, output, lang_code)`` for OCR training data.\nDefaults to false. No archival code wired in step 2c \u2014 the field is\nbaked into the contract so consumers don't need to upgrade their\nrequest shapes when archival turns on later."
      },
      "GenerateAssResponse": {
        "properties": {
          "file_id": {
            "type": "string",
            "title": "File Id"
          }
        },
        "type": "object",
        "required": [
          "file_id"
        ],
        "title": "GenerateAssResponse"
      },
      "GeneratePgsRequest": {
        "properties": {
          "native_file_id": {
            "type": "string",
            "title": "Native File Id"
          },
          "target_file_id": {
            "type": "string",
            "title": "Target File Id"
          },
          "target_lang_code": {
            "type": "string",
            "title": "Target Lang Code"
          },
          "styles": {
            "$ref": "#/components/schemas/StyleConfig"
          },
          "offsets": {
            "$ref": "#/components/schemas/TimingOffsets",
            "default": {
              "bottom_ms": 0,
              "top_ms": 0
            }
          },
          "source_resolution": {
            "$ref": "#/components/schemas/Resolution",
            "default": {
              "width": 1920,
              "height": 1080
            }
          },
          "output_resolution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Resolution"
              },
              {
                "type": "null"
              }
            ]
          },
          "opt_in_training": {
            "type": "boolean",
            "title": "Opt In Training",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "native_file_id",
          "target_file_id",
          "target_lang_code",
          "styles"
        ],
        "title": "GeneratePgsRequest",
        "description": "Client \u2192 server request for PGS generation. Async \u2014 returns a job ID.\n\nPGS rasterization takes 30s\u201310min depending on event count and\ncanvas resolution. Poll GET /jobs/{id} for progress. When ``state``\ntransitions to ``completed``, ``result_file_id`` resolves to the\ngenerated .sup via GET /files/{id}."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobAccepted": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "ass",
              "pgs",
              "mux"
            ],
            "title": "Kind"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind"
        ],
        "title": "JobAccepted"
      },
...
      "Resolution": {
        "properties": {
          "width": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Width",
            "default": 1920
          },
          "height": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Height",
            "default": 1080
          }
        },
        "type": "object",
        "title": "Resolution"
      },
...
      "RomanizedLayerStyle": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "fontname": {
            "type": "string",
            "title": "Fontname",
            "default": "Times New Roman"
          },
          "fontsize": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Fontsize",
            "default": 30
          },
          "bold": {
            "type": "boolean",
            "title": "Bold",
            "default": false
          },
          "italic": {
            "type": "boolean",
            "title": "Italic",
            "default": true
          },
          "primarycolor": {
            "type": "string",
            "title": "Primarycolor",
            "default": "#FFFFFF"
          },
          "primary_opacity": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Primary Opacity",
            "default": 100
          },
          "outlinecolor": {
            "type": "string",
            "title": "Outlinecolor",
            "default": "#000000"
          },
          "outline_opacity": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Outline Opacity",
            "default": 100
          },
          "backcolor": {
            "type": "string",
            "title": "Backcolor",
            "default": "#000000"
          },
          "back_opacity": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Back Opacity",
            "default": 0
          },
          "outline_enabled": {
            "type": "boolean",
            "title": "Outline Enabled",
            "default": true
          },
          "outline": {
            "type": "number",
            "minimum": 0.0,
            "title": "Outline",
            "default": 1.5
          },
          "shadow_enabled": {
            "type": "boolean",
            "title": "Shadow Enabled",
            "default": false
          },
          "shadow": {
            "type": "number",
            "minimum": 0.0,
            "title": "Shadow",
            "default": 1.5
          },
          "background_enabled": {
            "type": "boolean",
            "title": "Background Enabled",
            "default": false
          },
          "glow_enabled": {
            "type": "boolean",
            "title": "Glow Enabled",
            "default": false
          },
          "glow_radius": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Glow Radius",
            "default": 5
          },
          "glow_color_hex": {
            "type": "string",
            "title": "Glow Color Hex",
            "default": "#FFFF00"
          },
          "alignment": {
            "type": "integer",
            "maximum": 9.0,
            "minimum": 1.0,
            "title": "Alignment",
            "default": 8
          },
          "marginl": {
            "type": "integer",
            "title": "Marginl",
            "default": 10
          },
          "marginr": {
            "type": "integer",
            "title": "Marginr",
            "default": 10
          },
          "marginv": {
            "type": "integer",
            "title": "Marginv",
            "default": 10
          },
          "long_vowel_mode": {
            "type": "string",
            "enum": [
              "macrons",
              "doubled",
              "unmarked"
            ],
            "title": "Long Vowel Mode",
            "default": "macrons"
          }
        },
        "type": "object",
        "title": "RomanizedLayerStyle"
      },
...
      "StyleConfig": {
        "properties": {
          "Bottom": {
            "$ref": "#/components/schemas/BottomLayerStyle"
          },
          "Top": {
            "$ref": "#/components/schemas/TopLayerStyle"
          },
          "Romanized": {
            "$ref": "#/components/schemas/RomanizedLayerStyle"
          },
          "Annotation": {
            "$ref": "#/components/schemas/AnnotationLayerStyle"
          },
          "vertical_offset": {
            "type": "integer",
            "title": "Vertical Offset",
            "default": 0
          },
          "annotation_gap": {
            "type": "integer",
            "title": "Annotation Gap",
            "default": 2
          },
          "romanized_gap": {
            "type": "integer",
            "title": "Romanized Gap",
            "default": 0
          },
          "script_display": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "original",
                  "simplified",
                  "traditional"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Script Display"
          }
        },
        "type": "object",
        "title": "StyleConfig",
        "description": "The full styles dict the engine consumes.\n\nMirrors the hybrid shape used in the codebase: layer dicts keyed by\n``\"Bottom\"``/``\"Top\"``/``\"Romanized\"``/``\"Annotation\"`` plus a few\ntop-level scalars. Field aliases preserve the layer-name casing so a\nJSON payload can use either the alias (``{\"Bottom\": {...}}``) or the\nsnake_case attribute (``{\"bottom\": {...}}``)."
      },
...
      "SuggestFilenameRequest": {
        "properties": {
          "ext": {
            "type": "string",
            "title": "Ext"
          },
          "video_file_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video File Id"
          },
          "native_lang_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Native Lang Code"
          },
          "target_lang_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Lang Code"
          },
          "phonetic_system": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phonetic System"
          },
          "include_annotations": {
            "type": "boolean",
            "title": "Include Annotations",
            "default": false
          },
          "include_romanization": {
            "type": "boolean",
            "title": "Include Romanization",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "ext"
        ],
        "title": "SuggestFilenameRequest",
        "description": "Request a sanitized, metadata-aware default filename for a Save\ndialog. All inputs are optional \u2014 missing fields just get skipped.\n\nMirrors the Streamlit flow (loom_app.py:1252): pulls media title +\nyear from the video (when a ``video_file_id`` is supplied) and folds\nin the native/target language codes plus annotation / romanization\nsystem names."
      },
      "SuggestFilenameResponse": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          }
        },
        "type": "object",
        "required": [
          "filename"
        ],
        "title": "SuggestFilenameResponse"
      },
      "TimingOffsets": {
        "properties": {
          "bottom_ms": {
            "type": "integer",
            "title": "Bottom Ms",
            "default": 0
          },
          "top_ms": {
            "type": "integer",
            "title": "Top Ms",
            "default": 0
          }
        },
        "type": "object",
        "title": "TimingOffsets"
      },
      "TopLayerStyle": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "fontname": {
            "type": "string",
            "title": "Fontname",
            "default": "Noto Sans"
          },
          "fontsize": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Fontsize",
            "default": 52
          },
          "bold": {
            "type": "boolean",
            "title": "Bold",
            "default": false
          },
          "italic": {
            "type": "boolean",
            "title": "Italic",
            "default": false
          },
          "primarycolor": {
            "type": "string",
            "title": "Primarycolor",
            "default": "#FFFFFF"
          },
          "primary_opacity": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Primary Opacity",
            "default": 100
          },
          "outlinecolor": {
            "type": "string",
            "title": "Outlinecolor",
            "default": "#000000"
          },
          "outline_opacity": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Outline Opacity",
            "default": 100
          },
          "backcolor": {
            "type": "string",
            "title": "Backcolor",
            "default": "#000000"
          },
          "back_opacity": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Back Opacity",
            "default": 0
          },
          "outline_enabled": {
            "type": "boolean",
            "title": "Outline Enabled",
            "default": true
          },
          "outline": {
            "type": "number",
            "minimum": 0.0,
            "title": "Outline",
            "default": 2.5
          },
          "shadow_enabled": {
            "type": "boolean",
            "title": "Shadow Enabled",
            "default": false
          },
          "shadow": {
            "type": "number",
            "minimum": 0.0,
            "title": "Shadow",
            "default": 1.5
          },
          "background_enabled": {
            "type": "boolean",
            "title": "Background Enabled",
            "default": false
          },
          "glow_enabled": {
            "type": "boolean",
            "title": "Glow Enabled",
            "default": false
          },
          "glow_radius": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Glow Radius",
            "default": 5
          },
          "glow_color_hex": {
            "type": "string",
            "title": "Glow Color Hex",
            "default": "#FFFF00"
          },
          "alignment": {
            "type": "integer",
            "maximum": 9.0,
            "minimum": 1.0,
            "title": "Alignment",
            "default": 8
          },
          "marginl": {
            "type": "integer",
            "title": "Marginl",
            "default": 10
          },
          "marginr": {
            "type": "integer",
            "title": "Marginr",
            "default": 10
          },
          "marginv": {
            "type": "integer",
            "title": "Marginv",
            "default": 90
          }
        },
        "type": "object",
        "title": "TopLayerStyle"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}